pyproject.toml
[project]
name = "kalshi-arb"
version = "0.1.0"
description = "Cross-Asset Basis Arbitrage Trading System for Kalshi prediction markets"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "databento>=0.22.0",
    "aiohttp>=3.9.0",
    "websockets>=12.0",
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "redis>=5.0.0",
    "prometheus-client>=0.19.0",
    "cryptography>=41.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
]

[project.scripts]
kalshi-arb = "src.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py311"
